:recycle: refactor conversion from timestamp on rust side
