+++ frb_dart/lib/src/helpers.dart
@@ -6,6 +6,8 @@ import 'package:flutter_rust_bridge/src/platform_independent.dart';
 import 'package:meta/meta.dart';
 export 'ffi.dart';
 
+/// borrowed from flutter foundation [kIsWeb](https://api.flutter.dev/flutter/foundation/kIsWeb-constant.html),
+/// but allows for using it in a Dart context alike
 const bool kIsWeb = identical(0, 0.0);
 
 /// Allow custom setup hooks before ffi can be executed.
